Bounce mechanic + Input System
Added in the bounce mechanic so the players can now bounce away from eachother once they hit eachother. Inplemented the new Input System so the game can now be controlled with either any gamepad or keyboard. Made a new project because of unsolvable errors in the previous version when inplementing the new Input System from Unity.