Tiledrop pattern
Added in a Tile Manager script that keeps track of how many tiles are currently active. The tiles all have a tile script that, once apon called, give the player feedback that the tile is about the fall. The tile manager keeps looping over the tiles in the list until there are no more tiles available.