Ranking system + rank display UI
Adjusted the ranking system so that when a player gets hit he gets deactivated, removed from the playerlist and added in the ranklist. If there is only one player left he automaticely gets added to the ranking list. Afterwards the rankinglist gets flipped so that the person that won gets displayed on top. Added in a DelayResetCount variable to adjust when the scene needs to be reloaded. Need to change this later to DelayNextSceneLoad variable and that the next scene gets loaded in.