Fixed Ready Up system + Added Rank scene
Fixed issue where, once the minigame scene got reloaded the controls weren't assigned to the correct player. This has now been fixed by removing the player input on the UI elements. Added in a ranking scene so when the game is over it switches to that scene for a few seconds before going back to the minigame to replay it again.